services/alert_engine_py/requirements.txt
sqlalchemy==2.0.23
paho-mqtt==2.0.0
python-dateutil==2.8.2
psycopg2-binary==2.9.9

psycopg2-binary==2.9.9
